Use next segment start to calculate stub's spare bytes
The assumption that segments would start on paragraph (16-byte) boundaries
did not hold for Ultima Underworld II, so that heuristic had to be
replaced.

Also, include spare bytes and space for additional procs in logged details
about each overlay stub segment.

+++ UltimaPatcher/src/main/java/net/johnglassmyer/ultimapatcher/UltimaPatcher.java
@@ -387,7 +387,8 @@ public class UltimaPatcher {
 		}
 		Collections.reverse(segments);
 
-		return new Executable(exePath, file.length(), mzHeader, loadModule, fbovHeader, segments);
+		return new Executable(
+				exePath, (int) file.length(), mzHeader, loadModule, fbovHeader, segments);
 	}
 
 	private static List<Integer> parseMzRelocationsInLoadModule(byte[] relocationTableBytes) {
